set the number of notifications with evry 10 seconds
